requirements.txt
fastapi==0.103.1
uvicorn
psycopg2-binary
requests
python-dotenv
beautifulsoup4
lxml
html5lib
supabase
python-jose[cryptography]
google-generativeai
python-dateutil
numpy
duckduckgo-search
scikit-learn

# NOTE: keep Vercel serverless bundle small.
# Selenium + undetected-chromedriver are used for local ingestion only.
# Install them via requirements-worker.txt (or your local venv) instead.

